requirements.txt
Faker==21.0.0
psycopg2-binary==2.9.9
python-dateutil==2.8.2
six==1.16.0
